refactor: Moved code to separate class
