+++ src/main/java/bjs2_57242/Main.java
@@ -29,40 +29,12 @@ public class Main {
         showMatrix(matrix);
 
         System.out.println("Vertical transformation");
-        int[][] verticalTransformation = flipMatrix(matrix, FlipDirection.VERTICAL);
+        int[][] verticalTransformation = MatrixUtils.flipMatrix(matrix, FlipDirection.VERTICAL);
         showMatrix(verticalTransformation);
 
         System.out.println("Horizontal transformation");
-        int[][] horizontalTransformation = flipMatrix(matrix, FlipDirection.HORIZONTAL);
+        int[][] horizontalTransformation = MatrixUtils.flipMatrix(matrix, FlipDirection.HORIZONTAL);
         showMatrix(horizontalTransformation);
-    }
-
-    private static int[][] transformMatrix(int[][] matrix, MatrixTransformer transformer) {
-        int rowCount = matrix.length;
-        int columnCount = matrix[0].length;
-        int[][] result = new int[rowCount][columnCount];
-
-        for (int row = 0; row < rowCount; row++) {
-            for (int column = 0; column < columnCount; column++) {
-                Coordinates newCoordinates = transformer.transform(row, column, rowCount, columnCount);
-                result[newCoordinates.x()][newCoordinates.y()] = matrix[row][column];
-            }
-        }
-
-        return result;
-    }
-
-    public static int[][] flipMatrix(int[][] matrix, FlipDirection flipDirection) {
-        if (matrix == null) {
-            throw new IllegalArgumentException("Аргумент 'matrix' не может быть равен null");
-        }
-
-        MatrixTransformer transformer = switch (flipDirection) {
-            case VERTICAL -> (x, y, rowCount, columnCount) -> new Coordinates(rowCount - 1 - x, y);
-            case HORIZONTAL -> (x, y, rowCount, columnCount) -> new Coordinates(x, columnCount - 1 - y);
-        };
-
-        return transformMatrix(matrix, transformer);
     }
 
     private static void showMatrix(int[][] matrix) {
