+++ src/main/java/bjs2_57242/MatrixUtils.java
@@ -1,0 +1,31 @@
+package bjs2_57242;
+
+public class MatrixUtils {
+    private static int[][] transformMatrix(int[][] matrix, MatrixTransformer transformer) {
+        int rowCount = matrix.length;
+        int columnCount = matrix[0].length;
+        int[][] result = new int[rowCount][columnCount];
+
+        for (int row = 0; row < rowCount; row++) {
+            for (int column = 0; column < columnCount; column++) {
+                Coordinates newCoordinates = transformer.transform(row, column, rowCount, columnCount);
+                result[newCoordinates.x()][newCoordinates.y()] = matrix[row][column];
+            }
+        }
+
+        return result;
+    }
+
+    public static int[][] flipMatrix(int[][] matrix, FlipDirection flipDirection) {
+        if (matrix == null) {
+            throw new IllegalArgumentException("Аргумент 'matrix' не может быть равен null");
+        }
+
+        MatrixTransformer transformer = switch (flipDirection) {
+            case VERTICAL -> (x, y, rowCount, columnCount) -> new Coordinates(rowCount - 1 - x, y);
+            case HORIZONTAL -> (x, y, rowCount, columnCount) -> new Coordinates(x, columnCount - 1 - y);
+        };
+
+        return transformMatrix(matrix, transformer);
+    }
+}
